requirements.txt
Django==3.1.1
drf-jwt==1.17.2
django-cors-headers==3.5.0
requests==2.24.0
djangorestframework==3.12.1
cryptography==3.1.1
PyJWT==1.7.1
psycopg2-binary==2.8.6